GMT_DOS-ACTORS-CLIENTS_INTERFACE CODEC & FILING
2 new traits:

 * Codec: encoding/decoding object to writer/reader
 * Filing: encoding/decoding to file
