+++ interface/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "gmt_dos-actors-clients_interface"
-version = "1.1.0"
+version = "1.2.0"
 authors.workspace = true
 edition.workspace = true
 license.workspace = true
@@ -20,4 +20,7 @@ thiserror.workspace = true
 bincode = { workspace = true, optional = true }
 
 [features]
-load = ["serde", "bincode/serde"]
+filing = ["serde", "bincode/serde"]
+
+[package.metadata.docs.rs]
+all-features = true
